requirements.txt
PyQt6>=6.6.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
ffmpeg-python>=0.2.0
pyinstaller>=6.3.0
